Fix the permission sets. Also refactor the profiles tests a bit.
